Add component to navigation
