+++ package.json
@@ -16,7 +16,6 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
     "react-redux": "^8.0.5",
-    "react-router": "^6.4.5",
     "react-router-dom": "^6.4.4",
     "react-scripts": "5.0.1",
     "redux": "^4.2.0",
