+++ src/components/App.js
@@ -3,6 +3,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Aeroplanes from './Aeroplanes';
 import Aeroplane from './Aeroplane';
 import Navbar from './Navbar/Navbar';
+import ReservationFrom from './MakeReservation';
 
 function App() {
   return (
@@ -13,6 +14,7 @@ function App() {
           <Route path="/" element={<Aeroplanes />} />
           <Route path="/Aeroplanes" element={<Aeroplanes />} />
           <Route path="/Aeroplanes/:aeroplaneId" element={<Aeroplane />} />
+          <Route path="/MakeReservation" element={<ReservationFrom />} />
         </Routes>
       </div>
     </BrowserRouter>
