+++ src/components/Navbar/NavLinks.js
@@ -21,7 +21,7 @@ const NavLinks = ({ isOpen, setIsOpen }) => (
     </li>
     <li>
       <NavLink
-        to="/make-reservation"
+        to="/makeReservation"
         className="link"
         onClick={() => setIsOpen(!isOpen)}
       >
